+++ app/src/main/java/com/ildango/capstone/data/model/MyAlarmItemModel.kt
@@ -1,4 +1,4 @@
-package com.ildango.capstone.mypages.myalarmlist
+package com.ildango.capstone.data.model
 
 class MyAlarmItemModel:ArrayList<MyAlarmItem>()
