+++ app/src/main/java/com/ildango/capstone/data/repository/MyAlarmListRepository.kt
@@ -1,6 +1,6 @@
 package com.ildango.capstone.data.repository
 
-import com.ildango.capstone.mypages.myalarmlist.MyAlarmItem
+import com.ildango.capstone.data.model.MyAlarmItem
 import com.ildango.capstone.data.service.RetrofitClient
 import retrofit2.Response
