+++ app/src/main/java/com/ildango/capstone/data/repository/ProductRepository.kt
@@ -1,6 +1,6 @@
 package com.ildango.capstone.data.repository
 
-import com.ildango.capstone.resultdetail.ProductItemList
+import com.ildango.capstone.data.model.ProductItemList
 import com.ildango.capstone.data.service.RetrofitClient
 import retrofit2.Response
 import java.lang.Exception
