+++ app/src/main/java/com/ildango/capstone/data/service/MyAlarmService.kt
@@ -1,10 +1,8 @@
 package com.ildango.capstone.data.service
 
-import com.ildango.capstone.mypages.myalarmlist.MyAlarmItem
+import com.ildango.capstone.data.model.MyAlarmItem
 import retrofit2.Response
-import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.POST
 
 interface MyAlarmService {
     @GET("pricealarm/21")
