+++ app/src/main/java/com/ildango/capstone/data/service/MyWishService.kt
@@ -1,7 +1,7 @@
 package com.ildango.capstone.data.service
 
-import com.ildango.capstone.mypages.mywishlist.MyWishItem
-import com.ildango.capstone.mypages.mywishlist.MyWishPostItem
+import com.ildango.capstone.data.model.MyWishItem
+import com.ildango.capstone.data.model.MyWishPostItem
 import retrofit2.Response
 import retrofit2.http.*
 
@@ -10,7 +10,7 @@ interface MyWishService {
     suspend fun getAllWishProduct() : Response<List<MyWishItem>>
 
     @POST("myposts/add")
-    suspend fun setMyPost(@Body item:MyWishPostItem): Response<Int>
+    suspend fun setMyPost(@Body item: MyWishPostItem): Response<Int>
 
     @GET("myposts/exist")   /// myposts/exist?userId=3&postId=4
     suspend fun isItemExistInMyPosts(
