+++ app/src/main/java/com/ildango/capstone/data/service/RetrofitClient.kt
@@ -1,9 +1,7 @@
 package com.ildango.capstone.data.service
 
 import com.google.gson.GsonBuilder
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
