+++ app/src/main/java/com/ildango/capstone/mypages/myalarmlist/MyAlarmListViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
+import com.ildango.capstone.data.model.MyAlarmItem
 import com.ildango.capstone.data.repository.MyAlarmListRepository
 import kotlinx.coroutines.launch
 import retrofit2.Response
